Lua: Backend support for handling difficult memory relationship cases.
The memory problem is kind of a reference/ownership problem
where there is an implicit relationship between two objects in garbage
collected languages such as Lua and JavaScript.

Here's a more concrete example:

In SDL, if I create a renderer (SDL_CreateRenderer) and then create a
texture (SDL_CreateTextureFromSurface(renderer, bmp_surface)), the
texture is implicitly connected to the renderer, and the renderer must
stay alive longer than the texture.

Additionally, in garbage collection, the order of destruction of objects isn't
guaranteed. So if the renderer gets collected before the textures, bad
things happen.

Similarly, in Chipmunk, if I create a body (cpBodyNew) and add it to a
space aka world (cpSpaceNew), but don't hang onto the cpBody reference
explicitly, the body will get collected while in use by the cpSpace
and bad things happen.

So I need a way to express this relationship and keep a
reference between the objects. I've begun to see that this is both a
common pattern, and I think there may be a generalized way to solve
it. Since it occurs so much, I think it is better if SWIG can
automatically handle this for me if I can provide the correct
annotations.

This commit adds the backend support needed for Lua.
I have a new private field (in userdata the metatable as the path
of least resistance/impedance) which is an array which can hold strong
references to other objects. This allows the above relationships to be
expressed and inform the garbage collector what's going on. I have two
helper functions that can set an association and remove an
association.

SWIG_Lua_AssociateReference
SWIG_Lua_UnassociateReference

SWIG_Lua_AssociateReference:
For objectA and objectB, make objectB hold a reference to objectA, such that objectA cannot be collected until objectB is discarded.
Example: Tree: objectB is a child node of objectA. This will prevent the root (objectA)
of the tree from being collected as long as objectB is in the system.
2 params are expected to be passed in. Function returns with the 2 params still on the stack.

 there should be 2 params passed in
  (2 or -1 aka top of stack) objectB
  (1 or -2) objectA

The next ideal step is to add a %feature or something that allows this to be automatically invoked with the correct annotation.
However, this backend support allows custom %inline definitions to be written today.