JavaScriptCore: Public init function must be marked for exporting, or when compiling as a dynamic library, there are no symbols found. Additionally, the public API function needs to return a value that is C89 compatible.
In Visual Studio when compiling as a .dll, the missing SWIGEXPORT will cause no .lib file to be created which means you can't build/link any programs against the .dll that gets generated.

bool is C99 and technically not available on C89 (Visual Studio). It appears JavaScriptCore is providing some typedefs for bool because they use it in their own APIs. However, exporting this as a public API function seems like a bad idea so the value has been changed to int.